Move onInstruction function
